+++ src/OpenTelemetry.Exporter.Prometheus.HttpListener/PrometheusHttpListener.cs
@@ -129,10 +129,10 @@ internal sealed class PrometheusHttpListener : IDisposable
         }
     }
 
-    private static PrometheusProtocol Negotiate(HttpListenerRequest request, EscapingScheme defaultEscaping)
+    private static PrometheusProtocol Negotiate(HttpListenerRequest request)
     {
         var acceptHeader = request.Headers["Accept"];
-        return PrometheusHeadersParser.Negotiate(acceptHeader, defaultEscaping);
+        return PrometheusHeadersParser.Negotiate(acceptHeader);
     }
 
     /// <summary>
@@ -267,7 +267,7 @@ internal sealed class PrometheusHttpListener : IDisposable
 
             using var linkedCts = CancellationTokenSource.CreateLinkedTokenSource(requestCancelled.Token, cancellationToken);
 
-            var protocol = Negotiate(context.Request, this.exporter.DefaultEscapingScheme);
+            var protocol = Negotiate(context.Request);
 
             var collectionResponse = await this.exporter.CollectionManager.EnterCollect(protocol).ConfigureAwait(false);